feat(model): users を追加した
バックエンドに倣って auth.ts から UserResponse への移動も含む
