+++ src/model/auth.ts
@@ -1,3 +1,5 @@
+import { UserResponse } from './users'
+
 export interface SigninPayload {
   username: string
   password: string
@@ -6,13 +8,4 @@ export interface SigninPayload {
 export interface SigninResponse {
   user: UserResponse
   token: string
-}
-
-export interface UserResponse {
-  id: number
-  username: string
-  encryptedPassword: string
-  displayName?: string
-  createdAt: Date
-  updatedAt?: Date
 }
