+++ src/model/users.ts
@@ -1,0 +1,14 @@
+export interface UserPayload {
+  username: string
+  password: string
+  displayName?: string
+}
+
+export interface UserResponse {
+  id: number
+  username: string
+  encryptedPassword: string
+  displayName?: string
+  createdAt: Date
+  updatedAt?: Date
+}
